Enables optional images, assets & simple properties.
This change make sure that all values that are contained in the schema of a collection are present on the graphql object and can be queried. For object values (image, assets) it also makes sure that the localFile content is created so that childImageSharp can be used in the graphql query without a runtime error.

As a result there is some poisoning of the graphql data as some temporary values that are not contained in the graphql schema need to be added. To remove this data again after loading it is necessary to use a hoc in the page components that will clean up the data again.